feat(michelson-encoder): Better support for various tokens
Add support for more tokens and improved tests coverage
