+++ packages/tezos-ts-michelson-encoder/data/sample6.ts
@@ -1,5 +1,9 @@
-export const rpcContractResponse = {"manager":"tz1Ke2h7sDdakHJQh8WX4Z372du1KChsksyU","balance":"9354941173593","spendable":false,"delegate":{"setable":false,"value":"tz3WMqdzXqRWXwyvj5Hp2H7QEepaUuS7vd9K"},"script":{"code":[{"prim":"parameter","args":[{"prim":"or","args":[{"prim":"pair","args":[{"prim":"or","args":[{"prim":"or","args":[{"prim":"pair","args":[{"prim":"contract","args":[{"prim":"unit"}],"annots":["%dest"]},{"prim":"mutez","annots":["%transfer_amount"]}],"annots":["%Transfer"]},{"prim":"option","args":[{"prim":"pair","args":[{"prim":"contract","args":[{"prim":"unit"}],"annots":["%pour_dest"]},{"prim":"key","annots":["%pour_authorizer"]}]}],"annots":["%Set_pour"]}]},{"prim":"or","args":[{"prim":"pair","args":[{"prim":"list","args":[{"prim":"pair","args":[{"prim":"list","args":[{"prim":"key"}],"annots":["%signatories"]},{"prim":"nat","annots":["%group_threshold"]}]}],"annots":["%key_groups"]},{"prim":"nat","annots":["%overall_threshold"]}],"annots":["%Set_keys"]},{"prim":"option","args":[{"prim":"key_hash","annots":["%new_delegate"]}],"annots":["%Set_delegate"]}]}],"annots":["%action_input"]},{"prim":"list","args":[{"prim":"list","args":[{"prim":"option","args":[{"prim":"signature"}]}]}],"annots":["%signatures"]}],"annots":["%Action"]},{"prim":"option","args":[{"prim":"pair","args":[{"prim":"signature","annots":["%pour_auth"]},{"prim":"mutez","annots":["%pour_amount"]}],"annots":["%Pour"]}]}]}]},{"prim":"storage","args":[{"prim":"pair","args":[{"prim":"pair","args":[{"prim":"nat","annots":["%replay_counter"]},{"prim":"pair","args":[{"prim":"list","args":[{"prim":"pair","args":[{"prim":"list","args":[{"prim":"key"}],"annots":["%signatories"]},{"prim":"nat","annots":["%group_threshold"]}]}],"annots":["%key_groups"]},{"prim":"nat","annots":["%overall_threshold"]}],"annots":["%key_info"]}]},{"prim":"pair","args":[{"prim":"pair","args":[{"prim":"pair","args":[{"prim":"mutez","annots":["%vested_balance"]},{"prim":"mutez","annots":["%vesting_increment"]}],"annots":["%vesting_quantities"]},{"prim":"pair","args":[{"prim":"timestamp","annots":["%next_payout"]},{"prim":"int","annots":["%payout_interval"]}],"annots":["%vesting_schedule"]}],"annots":["%vesting"]},{"prim":"option","args":[{"prim":"pair","args":[{"prim":"contract","args":[{"prim":"unit"}],"annots":["%pour_dest"]},{"prim":"key","annots":["%pour_authorizer"]}]}],"annots":["%pour_info"]}]}]}]},{"prim":"code","args":[[{"prim":"DUP"},{"prim":"CAR"},{"prim":"IF_LEFT","args":[[{"prim":"DIP","args":[[{"prim":"CDR"},{"prim":"DUP"},{"prim":"NIL","args":[{"prim":"operation"}],"annots":["@operations"]},{"prim":"SWAP"}]]},{"prim":"SWAP"},{"prim":"CAR"},{"prim":"DUP"},{"prim":"CDR","annots":["%key_info","@%"]},{"prim":"DIP","args":[[{"prim":"CAR","annots":["%replay_counter","@%"]},{"prim":"DIP","args":[[{"prim":"DUP"},{"prim":"CDR","annots":["%signatures","@%"]}]]},{"prim":"SWAP"},{"prim":"DIP","args":[[{"prim":"DIP","args":[[{"prim":"CAR","annots":["%action_input","@%"]},{"prim":"DUP"}]]},{"prim":"SELF"},{"prim":"ADDRESS"},{"prim":"PAIR"},{"prim":"SWAP"},{"prim":"PAIR"},{"prim":"PACK","annots":["@packed"]}]]}]]},{"prim":"SWAP"},{"prim":"DIP","args":[[{"prim":"DUP"},{"prim":"CAR","annots":["%key_groups","@%"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%overall_threshold","@%"]},{"prim":"INT","annots":["@overall_counter"]},{"prim":"SWAP"}]]}]]},{"prim":"ITER","args":[[{"prim":"DIP","args":[[{"prim":"IF_CONS","args":[[{"prim":"DIP","args":[[{"prim":"SWAP"}]]},{"prim":"DUP"},{"prim":"CAR","annots":["@%%"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["@%%"]},{"prim":"INT","annots":["@group_counter"]},{"prim":"SWAP"}]]}],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}]]},{"prim":"ITER","args":[[{"prim":"DIP","args":[[{"prim":"IF_CONS","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}]]},{"prim":"IF_NONE","args":[[{"prim":"DROP"}],[{"prim":"SWAP"},{"prim":"DIP","args":[[{"prim":"DIP","args":[[{"prim":"SWAP"},{"prim":"DIP","args":[[{"prim":"DIP","args":[[{"prim":"PUSH","args":[{"prim":"int"},{"int":"-1"}]},{"prim":"ADD","annots":["@group_counter"]}]]}]]},{"prim":"DUP"}]]}]]},{"prim":"CHECK_SIGNATURE","annots":["@sig_ok"]},[{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}],{"prim":"SWAP"}]]}]]},[{"prim":"DIP","args":[[{"prim":"DIP","args":[[{"prim":"PUSH","args":[{"prim":"int"},{"int":"0"}]},[{"prim":"COMPARE"},{"prim":"GE","annots":["@group_threshold_met"]}],{"prim":"IF","args":[[{"prim":"DIP","args":[[{"prim":"PUSH","args":[{"prim":"int"},{"int":"-1"}]},{"prim":"ADD","annots":["@overall_counter"]}]]}],[]]}]]}]]}],{"prim":"IF_CONS","args":[[[{"prim":"UNIT"},{"prim":"FAILWITH"}]],[]]},{"prim":"SWAP"}]]},{"prim":"IF_CONS","args":[[[{"prim":"UNIT"},{"prim":"FAILWITH"}]],[]]},{"prim":"DROP"},{"prim":"PUSH","args":[{"prim":"int"},{"int":"0"}]},[{"prim":"COMPARE"},{"prim":"GE","annots":["@overall_threshold_met"]}],[{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}],[{"prim":"DIP","args":[[{"prim":"DIP","args":[[{"prim":"DUP"},{"prim":"CAR"},{"prim":"DUP"},{"prim":"CAR","annots":["%replay_counter","@%"]},{"prim":"PUSH","args":[{"prim":"nat"},{"int":"1"}]},{"prim":"ADD","annots":["@replay_counter"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%key_info","@%"]}]]},{"prim":"PAIR","annots":["%@","%@"]},{"prim":"DIP","args":[[{"prim":"CDR"}]]},{"prim":"PAIR"}]]}]]}],{"prim":"IF_LEFT","args":[[{"prim":"IF_LEFT","args":[[{"prim":"DUP"},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%transfer_amount","@%"]},{"prim":"DUP"},{"prim":"BALANCE","annots":["@balance"]},[[{"prim":"COMPARE"},{"prim":"GE"}],{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}],{"prim":"SWAP"},{"prim":"DIP","args":[[{"prim":"DIP","args":[[{"prim":"DUP"},{"prim":"CDR"},{"prim":"DUP"},{"prim":"CAR"},{"prim":"CAR","annots":["%vesting_quantities","@%"]},{"prim":"DUP"},{"prim":"CAR","annots":["%vested_balance","@%"]},{"prim":"DUP"}]]},{"prim":"DUP"},{"prim":"DIP","args":[[[[{"prim":"COMPARE"},{"prim":"LE"}],{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}]]]},{"prim":"SWAP"},{"prim":"SUB","annots":["@vested_balance"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%vesting_increment","@%"]}]]},{"prim":"PAIR","annots":["@vesting_quantities","%@","%@"]},{"prim":"DIP","args":[[{"prim":"DUP"},{"prim":"CAR"},{"prim":"CDR","annots":["%vesting_schedule","@%"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%pour_info","@%"]}]]}]]},{"prim":"PAIR","annots":["@vesting","%@","%@"]},{"prim":"PAIR","annots":["%@","%@"]},{"prim":"SWAP"},{"prim":"CAR"},{"prim":"PAIR"}]]}]]},{"prim":"DUP"},{"prim":"CDR","annots":["%transfer_amount","@%"]},{"prim":"DIP","args":[[{"prim":"CAR","annots":["%dest","@%"]}]]},{"prim":"UNIT"},{"prim":"TRANSFER_TOKENS","annots":["@transfer_op"]},{"prim":"CONS","annots":["@operations"]}],[{"prim":"SWAP"},{"prim":"DIP","args":[[{"prim":"SWAP"},{"prim":"DUP"},{"prim":"CAR"},{"prim":"DIP","args":[[{"prim":"CDR"},{"prim":"CAR","annots":["%vesting","@%"]},{"prim":"PAIR","annots":["%vesting","%pour_info"]}]]},{"prim":"PAIR","annots":["@storage"]}]]}]]}],[{"prim":"IF_LEFT","args":[[{"prim":"DUP"},{"prim":"DUP"},{"prim":"CAR","annots":["%key_groups","@%"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%overall_threshold","@new_threshold"]},{"prim":"DUP"},{"prim":"PUSH","args":[{"prim":"nat"},{"int":"0"}]},[[{"prim":"COMPARE"},{"prim":"LT"}],{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}],{"prim":"PUSH","args":[{"prim":"nat"},{"int":"0"}],"annots":["@counter"]}]]},{"prim":"ITER","args":[[{"prim":"DUP"},{"prim":"CAR","annots":["%signatories","@%%"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%group_threshold","@%%"]},{"prim":"DUP"},{"prim":"PUSH","args":[{"prim":"nat"},{"int":"0"}]},[[{"prim":"COMPARE"},{"prim":"LT"}],{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}]]]},{"prim":"SIZE","annots":["@num_keys"]},[[{"prim":"COMPARE"},{"prim":"GE"}],{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}],{"prim":"PUSH","args":[{"prim":"nat"},{"int":"1"}]},{"prim":"ADD","annots":["@counter"]}]]},[[{"prim":"COMPARE"},{"prim":"GE"}],{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}],{"prim":"SWAP"},{"prim":"DIP","args":[[{"prim":"DIP","args":[[{"prim":"DUP"},{"prim":"CAR"},{"prim":"CAR","annots":["%replay_counter"]}]]},{"prim":"SWAP"},{"prim":"PAIR","annots":["%@","%key_info"]},{"prim":"DIP","args":[[{"prim":"CDR"}]]},{"prim":"PAIR","annots":["@storage"]}]]}],[{"prim":"SET_DELEGATE","annots":["@set_delegate_op"]},{"prim":"CONS"}]]}]]},{"prim":"PAIR"}],[{"prim":"IF_NONE","args":[[{"prim":"CDR"},{"prim":"DUP"},{"prim":"CDR"},{"prim":"DUP"},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%pour_info","@%"]}]]},{"prim":"CAR","annots":["%vesting","@%"]},{"prim":"DUP"},{"prim":"CDR","annots":["%vesting_schedule","@%"]},{"prim":"DUP"},{"prim":"CAR","annots":["%next_payout","@%"]},{"prim":"DUP"},{"prim":"NOW"},[{"prim":"COMPARE"},{"prim":"GT"}],[{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}],{"prim":"DIP","args":[[{"prim":"CDR","annots":["%payout_interval","@%"]},{"prim":"DUP"}]]},{"prim":"ADD","annots":["@new_next_payout"]},{"prim":"PAIR","annots":["%next_payout","%@","@new_vesting_schedule"]},{"prim":"SWAP"},{"prim":"CAR","annots":["%vesting_quantities","@%"]},{"prim":"DUP"},{"prim":"CAR","annots":["%vested_balance","@%"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%vesting_increment","@%"]},{"prim":"DUP"}]]},{"prim":"ADD","annots":["@new_balance"]},{"prim":"PAIR","annots":["%vested_balance","%@"]},{"prim":"PAIR","annots":["%vesting_quantities","%vesting_schedule"]},{"prim":"PAIR","annots":["%vesting","%@"]},{"prim":"SWAP"},{"prim":"CAR"},{"prim":"PAIR","annots":["@storage"]},{"prim":"NIL","args":[{"prim":"operation"}]}],[{"prim":"DIP","args":[[{"prim":"CDR"},{"prim":"DUP"},{"prim":"CDR"},{"prim":"DUP"},{"prim":"CAR","annots":["%vesting","@%"]},{"prim":"DUP"},{"prim":"CAR","annots":["%vesting_quantities","@%"]},{"prim":"DUP"},{"prim":"CAR","annots":["%vested_balance","@%"]},{"prim":"DUP"}]]},{"prim":"DUP"},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%pour_amount","@%%"]},{"prim":"DUP"},{"prim":"DIP","args":[[[[{"prim":"COMPARE"},{"prim":"LE"}],{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}]]]},{"prim":"DUP"},{"prim":"BALANCE","annots":["@balance"]},[[{"prim":"COMPARE"},{"prim":"GE"}],{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}],{"prim":"SWAP"},{"prim":"SUB","annots":["@new_cur_balance"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%vesting_increment","@%"]}]]},{"prim":"PAIR","annots":["%vested_balance","%@","@vesting_quantities"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%vesting_schedule","@%"]}]]},{"prim":"PAIR","annots":["%@","%@","@vesting"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%pour_info","@%"]},{"prim":"DUP"},{"prim":"IF_NONE","args":[[[{"prim":"UNIT"},{"prim":"FAILWITH"}]],[]]}]]},{"prim":"SWAP"},{"prim":"DIP","args":[[{"prim":"PAIR","annots":["%@","%@"]},{"prim":"SWAP"},{"prim":"CAR"},{"prim":"DUP"},{"prim":"CAR","annots":["%replay_counter","@%"]},{"prim":"DUP"},{"prim":"PUSH","args":[{"prim":"nat"},{"int":"1"}]},{"prim":"ADD","annots":["@replay_counter"]},{"prim":"SWAP"},{"prim":"SELF"},{"prim":"ADDRESS"},{"prim":"PAIR"},{"prim":"DIP","args":[[{"prim":"DIP","args":[[{"prim":"CDR","annots":["%key_info","@%"]}]]},{"prim":"PAIR","annots":["%@","%@"]},{"prim":"PAIR","annots":["@storage"]}]]}]]}]]},{"prim":"DUP"},{"prim":"DIP","args":[[{"prim":"SWAP"},{"prim":"DUP"},{"prim":"CAR","annots":["%pour_dest","@%"]},{"prim":"DIP","args":[[{"prim":"DUP"},{"prim":"CDR","annots":["%pour_authorizer","@%"]},{"prim":"DIP","args":[[{"prim":"CAR","annots":["%pour_dest","@%"]},{"prim":"SWAP"},{"prim":"DUP"},{"prim":"CAR","annots":["%pour_auth","@%"]},{"prim":"DIP","args":[[{"prim":"CDR","annots":["%pour_amount","@%"]},{"prim":"SWAP"},{"prim":"PAIR"},{"prim":"PAIR"},{"prim":"PACK","annots":["@packed"]}]]}]]},{"prim":"CHECK_SIGNATURE","annots":["@pour_auth_ok"]},[{"prim":"IF","args":[[],[[{"prim":"UNIT"},{"prim":"FAILWITH"}]]]}],{"prim":"NIL","args":[{"prim":"operation"}]}]]}]]},{"prim":"CDR","annots":["%pour_amount","@%"]},{"prim":"UNIT"},{"prim":"TRANSFER_TOKENS","annots":["@pour_op"]},{"prim":"CONS"}]]},{"prim":"PAIR"}]]}]]}],"storage":{"prim":"Pair","args":[{"prim":"Pair","args":[{"int":"8001"},{"prim":"Pair","args":[[{"prim":"Pair","args":[[{"string":"edpkvS5QFv7KRGfa3b87gg9DBpxSm3NpSwnjhUjNBQrRUUR66F7C9g"},{"string":"edpktm3zeGMzfzFuqgyYftt7uNyVRANTjrJCdU7bURwgGb9bRZwmJq"},{"string":"edpkucCnbeGPWNvGHeTQ5hENHPrc8txBBiQXNphu3jgv9KYbhQBovd"},{"string":"edpkuNjKKT48xBoT5asPrWdmuM1Yw8D93MwgFgVvtca8jb5pstzaCh"}],{"int":"2"}]},{"prim":"Pair","args":[[{"string":"edpkukogJzCZSAfc5foRpEZGsryMjiXj77VxcFLzZoreB1vZWFeKc2"},{"string":"edpkuSZ1GoM6MALh4fPZBrwDhGwY9vENEoyctcXuDK3yoiX4xWhMaA"},{"string":"edpkvXL9B32DcEbMNiMangcSFMvAd8NBwH8AfmRb6iBbHxLgx3J59P"},{"string":"edpkumCFhgS94cjZXiFnnq7MFsaWBz4Tp78AX2fZfhB9J9hcgcKxgy"}],{"int":"2"}]},{"prim":"Pair","args":[[{"string":"edpkv9eiqPP8bxCPEem6BBUNokwwy64G7ARYov77oNyCfK6fP3H5hz"},{"string":"edpktjDaR4gmEZNSnnuSWMmJ4fhYmMD4Lk1WkDCfWRWdbGHQ9eNT45"},{"string":"edpkuapDV7oaWqK5fStnPbhEm2g7kvhhtbBNs9hhq3KWXS2qcWcdNQ"},{"string":"edpkuzAzJ5yQeSHsfRqwkdGBELdJjAGGR6gw3Ar1iVVDvFef5e6Juk"}],{"int":"2"}]},{"prim":"Pair","args":[[{"string":"edpkuhBM7RhVwtityNcxMZchDX8yAQejxe8qhBLBLKxuNtDYaBto7u"},{"string":"edpkuALgGd9Y5SWLXPfUJV9EG2rv6EkCBA4C47mrt4cAmE11EHQrCs"},{"string":"edpkvZRBMfmC2pcv6xnbHPzKpwH7U9sPgFieXGjQFdW2Tzu9vcd5f8"},{"string":"edpkvZkwaGFpqU3aRuUedeGx1xah6wLMre4nNYMvNyvaDZgKicvFWe"}],{"int":"2"}]},{"prim":"Pair","args":[[{"string":"edpkuBBAoSWWAUfRNGVZu5H4TBnCmRVhWW1sZApLA53pYfY3tKjnev"},{"string":"edpkvQpud1huJmJqUfF84uHjkpR7AQ32uKP1wkEqnJhgH1nXPQheUh"},{"string":"edpkuiacWZrSDn6LtkXkwq9mY7VPefGzoyZoL1PWGBj4EngM95D1Md"},{"string":"edpkuXtt222Rdkg2dAQzcBkPL9EAn9e9RzjFSYr3MhG2NnjS7ihMxh"}],{"int":"2"}]},{"prim":"Pair","args":[[{"string":"edpkv7k8jhUUuxHp4b2rrg2o9mZZB2SxM8i4bh1goebvQEporktRrr"},{"string":"edpkvStZUa8YbsVuFQWHXWoBkJ8kyf4W26TXjRu2HXS7ymfeHrnquP"},{"string":"edpktxdeLBghuHVyAq8q7SUKEJNMeFwPKVtWyJv9Pi51jWPXwhiFU4"},{"string":"edpkvSC9HasAnmjfX3Qt5f88PVo24kx5LRzzpqNVzq5Q9Rm7Q71ubB"}],{"int":"2"}]},{"prim":"Pair","args":[[{"string":"edpkvZ471mmT7nCrge44mzLwW6mmpLh5n2ykXX5zHunqgxMc4JcTgE"},{"string":"edpktsM8qruHNaGKM5wor6UDmkn5YeH9rekmTayUMAP5zyL8RCvGic"},{"string":"edpkuT8WzhehsZjyUqCYVQEvwUJ4Nez2mSjfrWxPvnUNCZjjYHbyMB"},{"string":"edpktepTo31j8SYSBCupfRA8VEdTXXvgtbZEEejuumpAGU6xKEi4sf"}],{"int":"2"}]}],{"int":"4"}]}]},{"prim":"Pair","args":[{"prim":"Pair","args":[{"prim":"Pair","args":[{"int":"0"},{"int":"199041301565"}]},{"prim":"Pair","args":[{"string":"2018-07-30T05:30:00Z"},{"int":"2629800"}]}]},{"prim":"Some","args":[{"prim":"Pair","args":[{"string":"tz3WMqdzXqRWXwyvj5Hp2H7QEepaUuS7vd9K"},{"string":"edpkv4vUwGVVYnmmuafbEirXrXhT1bzcuJ2xQ3SHfeUVUr56YwgipC"}]}]}]}]}},"counter":"0"}
+export const rpcContractResponse = { "manager": "tz1Ke2h7sDdakHJQh8WX4Z372du1KChsksyU", "balance": "9354941173593", "spendable": false, "delegate": { "setable": false, "value": "tz3WMqdzXqRWXwyvj5Hp2H7QEepaUuS7vd9K" }, "script": { "code": [{ "prim": "parameter", "args": [{ "prim": "or", "args": [{ "prim": "pair", "args": [{ "prim": "or", "args": [{ "prim": "or", "args": [{ "prim": "pair", "args": [{ "prim": "contract", "args": [{ "prim": "unit" }], "annots": ["%dest"] }, { "prim": "mutez", "annots": ["%transfer_amount"] }], "annots": ["%Transfer"] }, { "prim": "option", "args": [{ "prim": "pair", "args": [{ "prim": "contract", "args": [{ "prim": "unit" }], "annots": ["%pour_dest"] }, { "prim": "key", "annots": ["%pour_authorizer"] }] }], "annots": ["%Set_pour"] }] }, { "prim": "or", "args": [{ "prim": "pair", "args": [{ "prim": "list", "args": [{ "prim": "pair", "args": [{ "prim": "list", "args": [{ "prim": "key" }], "annots": ["%signatories"] }, { "prim": "nat", "annots": ["%group_threshold"] }] }], "annots": ["%key_groups"] }, { "prim": "nat", "annots": ["%overall_threshold"] }], "annots": ["%Set_keys"] }, { "prim": "option", "args": [{ "prim": "key_hash", "annots": ["%new_delegate"] }], "annots": ["%Set_delegate"] }] }], "annots": ["%action_input"] }, { "prim": "list", "args": [{ "prim": "list", "args": [{ "prim": "option", "args": [{ "prim": "signature" }] }] }], "annots": ["%signatures"] }], "annots": ["%Action"] }, { "prim": "option", "args": [{ "prim": "pair", "args": [{ "prim": "signature", "annots": ["%pour_auth"] }, { "prim": "mutez", "annots": ["%pour_amount"] }], "annots": ["%Pour"] }] }] }] }, { "prim": "storage", "args": [{ "prim": "pair", "args": [{ "prim": "pair", "args": [{ "prim": "nat", "annots": ["%replay_counter"] }, { "prim": "pair", "args": [{ "prim": "list", "args": [{ "prim": "pair", "args": [{ "prim": "list", "args": [{ "prim": "key" }], "annots": ["%signatories"] }, { "prim": "nat", "annots": ["%group_threshold"] }] }], "annots": ["%key_groups"] }, { "prim": "nat", "annots": ["%overall_threshold"] }], "annots": ["%key_info"] }] }, { "prim": "pair", "args": [{ "prim": "pair", "args": [{ "prim": "pair", "args": [{ "prim": "mutez", "annots": ["%vested_balance"] }, { "prim": "mutez", "annots": ["%vesting_increment"] }], "annots": ["%vesting_quantities"] }, { "prim": "pair", "args": [{ "prim": "timestamp", "annots": ["%next_payout"] }, { "prim": "int", "annots": ["%payout_interval"] }], "annots": ["%vesting_schedule"] }], "annots": ["%vesting"] }, { "prim": "option", "args": [{ "prim": "pair", "args": [{ "prim": "contract", "args": [{ "prim": "unit" }], "annots": ["%pour_dest"] }, { "prim": "key", "annots": ["%pour_authorizer"] }] }], "annots": ["%pour_info"] }] }] }] }, { "prim": "code", "args": [[{ "prim": "DUP" }, { "prim": "CAR" }, { "prim": "IF_LEFT", "args": [[{ "prim": "DIP", "args": [[{ "prim": "CDR" }, { "prim": "DUP" }, { "prim": "NIL", "args": [{ "prim": "operation" }], "annots": ["@operations"] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }, { "prim": "CAR" }, { "prim": "DUP" }, { "prim": "CDR", "annots": ["%key_info", "@%"] }, { "prim": "DIP", "args": [[{ "prim": "CAR", "annots": ["%replay_counter", "@%"] }, { "prim": "DIP", "args": [[{ "prim": "DUP" }, { "prim": "CDR", "annots": ["%signatures", "@%"] }]] }, { "prim": "SWAP" }, { "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "CAR", "annots": ["%action_input", "@%"] }, { "prim": "DUP" }]] }, { "prim": "SELF" }, { "prim": "ADDRESS" }, { "prim": "PAIR" }, { "prim": "SWAP" }, { "prim": "PAIR" }, { "prim": "PACK", "annots": ["@packed"] }]] }]] }, { "prim": "SWAP" }, { "prim": "DIP", "args": [[{ "prim": "DUP" }, { "prim": "CAR", "annots": ["%key_groups", "@%"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%overall_threshold", "@%"] }, { "prim": "INT", "annots": ["@overall_counter"] }, { "prim": "SWAP" }]] }]] }, { "prim": "ITER", "args": [[{ "prim": "DIP", "args": [[{ "prim": "IF_CONS", "args": [[{ "prim": "DIP", "args": [[{ "prim": "SWAP" }]] }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["@%%"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["@%%"] }, { "prim": "INT", "annots": ["@group_counter"] }, { "prim": "SWAP" }]] }], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }]] }, { "prim": "ITER", "args": [[{ "prim": "DIP", "args": [[{ "prim": "IF_CONS", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }]] }, { "prim": "IF_NONE", "args": [[{ "prim": "DROP" }], [{ "prim": "SWAP" }, { "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "SWAP" }, { "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "PUSH", "args": [{ "prim": "int" }, { "int": "-1" }] }, { "prim": "ADD", "annots": ["@group_counter"] }]] }]] }, { "prim": "DUP" }]] }]] }, { "prim": "CHECK_SIGNATURE", "annots": ["@sig_ok"] }, [{ "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }], { "prim": "SWAP" }]] }]] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "PUSH", "args": [{ "prim": "int" }, { "int": "0" }] }, [{ "prim": "COMPARE" }, { "prim": "GE", "annots": ["@group_threshold_met"] }], { "prim": "IF", "args": [[{ "prim": "DIP", "args": [[{ "prim": "PUSH", "args": [{ "prim": "int" }, { "int": "-1" }] }, { "prim": "ADD", "annots": ["@overall_counter"] }]] }], []] }]] }]] }], { "prim": "IF_CONS", "args": [[[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]], []] }, { "prim": "SWAP" }]] }, { "prim": "IF_CONS", "args": [[[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]], []] }, { "prim": "DROP" }, { "prim": "PUSH", "args": [{ "prim": "int" }, { "int": "0" }] }, [{ "prim": "COMPARE" }, { "prim": "GE", "annots": ["@overall_threshold_met"] }], [{ "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }], [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }, { "prim": "CAR" }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["%replay_counter", "@%"] }, { "prim": "PUSH", "args": [{ "prim": "nat" }, { "int": "1" }] }, { "prim": "ADD", "annots": ["@replay_counter"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%key_info", "@%"] }]] }, { "prim": "PAIR", "annots": ["%@", "%@"] }, { "prim": "DIP", "args": [[{ "prim": "CDR" }]] }, { "prim": "PAIR" }]] }]] }], { "prim": "IF_LEFT", "args": [[{ "prim": "IF_LEFT", "args": [[{ "prim": "DUP" }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%transfer_amount", "@%"] }, { "prim": "DUP" }, { "prim": "BALANCE", "annots": ["@balance"] }, [[{ "prim": "COMPARE" }, { "prim": "GE" }], { "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }], { "prim": "SWAP" }, { "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }, { "prim": "CDR" }, { "prim": "DUP" }, { "prim": "CAR" }, { "prim": "CAR", "annots": ["%vesting_quantities", "@%"] }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["%vested_balance", "@%"] }, { "prim": "DUP" }]] }, { "prim": "DUP" }, { "prim": "DIP", "args": [[[[{ "prim": "COMPARE" }, { "prim": "LE" }], { "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }]]] }, { "prim": "SWAP" }, { "prim": "SUB", "annots": ["@vested_balance"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%vesting_increment", "@%"] }]] }, { "prim": "PAIR", "annots": ["@vesting_quantities", "%@", "%@"] }, { "prim": "DIP", "args": [[{ "prim": "DUP" }, { "prim": "CAR" }, { "prim": "CDR", "annots": ["%vesting_schedule", "@%"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%pour_info", "@%"] }]] }]] }, { "prim": "PAIR", "annots": ["@vesting", "%@", "%@"] }, { "prim": "PAIR", "annots": ["%@", "%@"] }, { "prim": "SWAP" }, { "prim": "CAR" }, { "prim": "PAIR" }]] }]] }, { "prim": "DUP" }, { "prim": "CDR", "annots": ["%transfer_amount", "@%"] }, { "prim": "DIP", "args": [[{ "prim": "CAR", "annots": ["%dest", "@%"] }]] }, { "prim": "UNIT" }, { "prim": "TRANSFER_TOKENS", "annots": ["@transfer_op"] }, { "prim": "CONS", "annots": ["@operations"] }], [{ "prim": "SWAP" }, { "prim": "DIP", "args": [[{ "prim": "SWAP" }, { "prim": "DUP" }, { "prim": "CAR" }, { "prim": "DIP", "args": [[{ "prim": "CDR" }, { "prim": "CAR", "annots": ["%vesting", "@%"] }, { "prim": "PAIR", "annots": ["%vesting", "%pour_info"] }]] }, { "prim": "PAIR", "annots": ["@storage"] }]] }]] }], [{ "prim": "IF_LEFT", "args": [[{ "prim": "DUP" }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["%key_groups", "@%"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%overall_threshold", "@new_threshold"] }, { "prim": "DUP" }, { "prim": "PUSH", "args": [{ "prim": "nat" }, { "int": "0" }] }, [[{ "prim": "COMPARE" }, { "prim": "LT" }], { "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }], { "prim": "PUSH", "args": [{ "prim": "nat" }, { "int": "0" }], "annots": ["@counter"] }]] }, { "prim": "ITER", "args": [[{ "prim": "DUP" }, { "prim": "CAR", "annots": ["%signatories", "@%%"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%group_threshold", "@%%"] }, { "prim": "DUP" }, { "prim": "PUSH", "args": [{ "prim": "nat" }, { "int": "0" }] }, [[{ "prim": "COMPARE" }, { "prim": "LT" }], { "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }]]] }, { "prim": "SIZE", "annots": ["@num_keys"] }, [[{ "prim": "COMPARE" }, { "prim": "GE" }], { "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }], { "prim": "PUSH", "args": [{ "prim": "nat" }, { "int": "1" }] }, { "prim": "ADD", "annots": ["@counter"] }]] }, [[{ "prim": "COMPARE" }, { "prim": "GE" }], { "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }], { "prim": "SWAP" }, { "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }, { "prim": "CAR" }, { "prim": "CAR", "annots": ["%replay_counter"] }]] }, { "prim": "SWAP" }, { "prim": "PAIR", "annots": ["%@", "%key_info"] }, { "prim": "DIP", "args": [[{ "prim": "CDR" }]] }, { "prim": "PAIR", "annots": ["@storage"] }]] }], [{ "prim": "SET_DELEGATE", "annots": ["@set_delegate_op"] }, { "prim": "CONS" }]] }]] }, { "prim": "PAIR" }], [{ "prim": "IF_NONE", "args": [[{ "prim": "CDR" }, { "prim": "DUP" }, { "prim": "CDR" }, { "prim": "DUP" }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%pour_info", "@%"] }]] }, { "prim": "CAR", "annots": ["%vesting", "@%"] }, { "prim": "DUP" }, { "prim": "CDR", "annots": ["%vesting_schedule", "@%"] }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["%next_payout", "@%"] }, { "prim": "DUP" }, { "prim": "NOW" }, [{ "prim": "COMPARE" }, { "prim": "GT" }], [{ "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }], { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%payout_interval", "@%"] }, { "prim": "DUP" }]] }, { "prim": "ADD", "annots": ["@new_next_payout"] }, { "prim": "PAIR", "annots": ["%next_payout", "%@", "@new_vesting_schedule"] }, { "prim": "SWAP" }, { "prim": "CAR", "annots": ["%vesting_quantities", "@%"] }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["%vested_balance", "@%"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%vesting_increment", "@%"] }, { "prim": "DUP" }]] }, { "prim": "ADD", "annots": ["@new_balance"] }, { "prim": "PAIR", "annots": ["%vested_balance", "%@"] }, { "prim": "PAIR", "annots": ["%vesting_quantities", "%vesting_schedule"] }, { "prim": "PAIR", "annots": ["%vesting", "%@"] }, { "prim": "SWAP" }, { "prim": "CAR" }, { "prim": "PAIR", "annots": ["@storage"] }, { "prim": "NIL", "args": [{ "prim": "operation" }] }], [{ "prim": "DIP", "args": [[{ "prim": "CDR" }, { "prim": "DUP" }, { "prim": "CDR" }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["%vesting", "@%"] }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["%vesting_quantities", "@%"] }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["%vested_balance", "@%"] }, { "prim": "DUP" }]] }, { "prim": "DUP" }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%pour_amount", "@%%"] }, { "prim": "DUP" }, { "prim": "DIP", "args": [[[[{ "prim": "COMPARE" }, { "prim": "LE" }], { "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }]]] }, { "prim": "DUP" }, { "prim": "BALANCE", "annots": ["@balance"] }, [[{ "prim": "COMPARE" }, { "prim": "GE" }], { "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }], { "prim": "SWAP" }, { "prim": "SUB", "annots": ["@new_cur_balance"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%vesting_increment", "@%"] }]] }, { "prim": "PAIR", "annots": ["%vested_balance", "%@", "@vesting_quantities"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%vesting_schedule", "@%"] }]] }, { "prim": "PAIR", "annots": ["%@", "%@", "@vesting"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%pour_info", "@%"] }, { "prim": "DUP" }, { "prim": "IF_NONE", "args": [[[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]], []] }]] }, { "prim": "SWAP" }, { "prim": "DIP", "args": [[{ "prim": "PAIR", "annots": ["%@", "%@"] }, { "prim": "SWAP" }, { "prim": "CAR" }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["%replay_counter", "@%"] }, { "prim": "DUP" }, { "prim": "PUSH", "args": [{ "prim": "nat" }, { "int": "1" }] }, { "prim": "ADD", "annots": ["@replay_counter"] }, { "prim": "SWAP" }, { "prim": "SELF" }, { "prim": "ADDRESS" }, { "prim": "PAIR" }, { "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%key_info", "@%"] }]] }, { "prim": "PAIR", "annots": ["%@", "%@"] }, { "prim": "PAIR", "annots": ["@storage"] }]] }]] }]] }, { "prim": "DUP" }, { "prim": "DIP", "args": [[{ "prim": "SWAP" }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["%pour_dest", "@%"] }, { "prim": "DIP", "args": [[{ "prim": "DUP" }, { "prim": "CDR", "annots": ["%pour_authorizer", "@%"] }, { "prim": "DIP", "args": [[{ "prim": "CAR", "annots": ["%pour_dest", "@%"] }, { "prim": "SWAP" }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["%pour_auth", "@%"] }, { "prim": "DIP", "args": [[{ "prim": "CDR", "annots": ["%pour_amount", "@%"] }, { "prim": "SWAP" }, { "prim": "PAIR" }, { "prim": "PAIR" }, { "prim": "PACK", "annots": ["@packed"] }]] }]] }, { "prim": "CHECK_SIGNATURE", "annots": ["@pour_auth_ok"] }, [{ "prim": "IF", "args": [[], [[{ "prim": "UNIT" }, { "prim": "FAILWITH" }]]] }], { "prim": "NIL", "args": [{ "prim": "operation" }] }]] }]] }, { "prim": "CDR", "annots": ["%pour_amount", "@%"] }, { "prim": "UNIT" }, { "prim": "TRANSFER_TOKENS", "annots": ["@pour_op"] }, { "prim": "CONS" }]] }, { "prim": "PAIR" }]] }]] }], "storage": { "prim": "Pair", "args": [{ "prim": "Pair", "args": [{ "int": "8001" }, { "prim": "Pair", "args": [[{ "prim": "Pair", "args": [[{ "string": "edpkvS5QFv7KRGfa3b87gg9DBpxSm3NpSwnjhUjNBQrRUUR66F7C9g" }, { "string": "edpktm3zeGMzfzFuqgyYftt7uNyVRANTjrJCdU7bURwgGb9bRZwmJq" }, { "string": "edpkucCnbeGPWNvGHeTQ5hENHPrc8txBBiQXNphu3jgv9KYbhQBovd" }, { "string": "edpkuNjKKT48xBoT5asPrWdmuM1Yw8D93MwgFgVvtca8jb5pstzaCh" }], { "int": "2" }] }, { "prim": "Pair", "args": [[{ "string": "edpkukogJzCZSAfc5foRpEZGsryMjiXj77VxcFLzZoreB1vZWFeKc2" }, { "string": "edpkuSZ1GoM6MALh4fPZBrwDhGwY9vENEoyctcXuDK3yoiX4xWhMaA" }, { "string": "edpkvXL9B32DcEbMNiMangcSFMvAd8NBwH8AfmRb6iBbHxLgx3J59P" }, { "string": "edpkumCFhgS94cjZXiFnnq7MFsaWBz4Tp78AX2fZfhB9J9hcgcKxgy" }], { "int": "2" }] }, { "prim": "Pair", "args": [[{ "string": "edpkv9eiqPP8bxCPEem6BBUNokwwy64G7ARYov77oNyCfK6fP3H5hz" }, { "string": "edpktjDaR4gmEZNSnnuSWMmJ4fhYmMD4Lk1WkDCfWRWdbGHQ9eNT45" }, { "string": "edpkuapDV7oaWqK5fStnPbhEm2g7kvhhtbBNs9hhq3KWXS2qcWcdNQ" }, { "string": "edpkuzAzJ5yQeSHsfRqwkdGBELdJjAGGR6gw3Ar1iVVDvFef5e6Juk" }], { "int": "2" }] }, { "prim": "Pair", "args": [[{ "string": "edpkuhBM7RhVwtityNcxMZchDX8yAQejxe8qhBLBLKxuNtDYaBto7u" }, { "string": "edpkuALgGd9Y5SWLXPfUJV9EG2rv6EkCBA4C47mrt4cAmE11EHQrCs" }, { "string": "edpkvZRBMfmC2pcv6xnbHPzKpwH7U9sPgFieXGjQFdW2Tzu9vcd5f8" }, { "string": "edpkvZkwaGFpqU3aRuUedeGx1xah6wLMre4nNYMvNyvaDZgKicvFWe" }], { "int": "2" }] }, { "prim": "Pair", "args": [[{ "string": "edpkuBBAoSWWAUfRNGVZu5H4TBnCmRVhWW1sZApLA53pYfY3tKjnev" }, { "string": "edpkvQpud1huJmJqUfF84uHjkpR7AQ32uKP1wkEqnJhgH1nXPQheUh" }, { "string": "edpkuiacWZrSDn6LtkXkwq9mY7VPefGzoyZoL1PWGBj4EngM95D1Md" }, { "string": "edpkuXtt222Rdkg2dAQzcBkPL9EAn9e9RzjFSYr3MhG2NnjS7ihMxh" }], { "int": "2" }] }, { "prim": "Pair", "args": [[{ "string": "edpkv7k8jhUUuxHp4b2rrg2o9mZZB2SxM8i4bh1goebvQEporktRrr" }, { "string": "edpkvStZUa8YbsVuFQWHXWoBkJ8kyf4W26TXjRu2HXS7ymfeHrnquP" }, { "string": "edpktxdeLBghuHVyAq8q7SUKEJNMeFwPKVtWyJv9Pi51jWPXwhiFU4" }, { "string": "edpkvSC9HasAnmjfX3Qt5f88PVo24kx5LRzzpqNVzq5Q9Rm7Q71ubB" }], { "int": "2" }] }, { "prim": "Pair", "args": [[{ "string": "edpkvZ471mmT7nCrge44mzLwW6mmpLh5n2ykXX5zHunqgxMc4JcTgE" }, { "string": "edpktsM8qruHNaGKM5wor6UDmkn5YeH9rekmTayUMAP5zyL8RCvGic" }, { "string": "edpkuT8WzhehsZjyUqCYVQEvwUJ4Nez2mSjfrWxPvnUNCZjjYHbyMB" }, { "string": "edpktepTo31j8SYSBCupfRA8VEdTXXvgtbZEEejuumpAGU6xKEi4sf" }], { "int": "2" }] }], { "int": "4" }] }] }, { "prim": "Pair", "args": [{ "prim": "Pair", "args": [{ "prim": "Pair", "args": [{ "int": "0" }, { "int": "199041301565" }] }, { "prim": "Pair", "args": [{ "string": "2018-07-30T05:30:00Z" }, { "int": "2629800" }] }] }, { "prim": "Some", "args": [{ "prim": "Pair", "args": [{ "string": "tz3WMqdzXqRWXwyvj5Hp2H7QEepaUuS7vd9K" }, { "string": "edpkv4vUwGVVYnmmuafbEirXrXhT1bzcuJ2xQ3SHfeUVUr56YwgipC" }] }] }] }] } }, "counter": "0" }
 
 export const storage = rpcContractResponse.script.code.find(
   x => x.prim === 'storage'
 )!.args[0] as any;
+
+export const params = rpcContractResponse.script.code.find(
+  x => x.prim === 'parameter'
+)!.args[0];
