+++ packages/tezos-ts-michelson-encoder/data/sample9.ts
@@ -1,0 +1,9 @@
+export const rpcContractResponse = { "manager": "tz1ZLyEcFfXyu75ASsjAt5hgaR4sU5PSbLkg", "balance": "0", "spendable": true, "delegate": { "setable": true }, "script": { "code": [{ "prim": "parameter", "args": [{ "prim": "or", "args": [{ "prim": "pair", "args": [{ "prim": "address" }, { "prim": "pair", "args": [{ "prim": "string" }, { "prim": "option", "args": [{ "prim": "bytes" }] }] }] }, { "prim": "or", "args": [{ "prim": "mutez" }, { "prim": "or", "args": [{ "prim": "pair", "args": [{ "prim": "address" }, { "prim": "bool" }] }, { "prim": "lambda", "args": [{ "prim": "pair", "args": [{ "prim": "address" }, { "prim": "pair", "args": [{ "prim": "string" }, { "prim": "option", "args": [{ "prim": "bytes" }] }] }] }, { "prim": "operation" }] }] }] }] }] }, { "prim": "storage", "args": [{ "prim": "pair", "args": [{ "prim": "big_map", "args": [{ "prim": "address" }, { "prim": "pair", "args": [{ "prim": "string" }, { "prim": "pair", "args": [{ "prim": "option", "args": [{ "prim": "bytes" }] }, { "prim": "pair", "args": [{ "prim": "mutez" }, { "prim": "pair", "args": [{ "prim": "or", "args": [{ "prim": "unit" }, { "prim": "or", "args": [{ "prim": "address" }, { "prim": "unit" }] }] }, { "prim": "timestamp" }] }] }] }] }] }, { "prim": "pair", "args": [{ "prim": "mutez" }, { "prim": "pair", "args": [{ "prim": "set", "args": [{ "prim": "address" }] }, { "prim": "option", "args": [{ "prim": "lambda", "args": [{ "prim": "pair", "args": [{ "prim": "address" }, { "prim": "pair", "args": [{ "prim": "string" }, { "prim": "option", "args": [{ "prim": "bytes" }] }] }] }, { "prim": "operation" }] }] }] }] }] }] }, { "prim": "code", "args": [[{ "prim": "DUP" }, { "prim": "DIP", "args": [[{ "prim": "CDR" }]] }, { "prim": "CAR" }, { "prim": "NIL", "args": [{ "prim": "operation" }], "annots": ["@return"] }, { "prim": "SENDER", "annots": ["@sender"] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "CDR" }, { "prim": "CDR" }, { "prim": "CAR" }], [{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@sender"] }]] }, { "prim": "SWAP" }], { "prim": "MEM", "annots": ["@owned"] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@parameter"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "IF_LEFT", "args": [[[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "CDR" }, { "prim": "CDR" }, { "prim": "CDR" }], { "prim": "IF_NONE", "args": [[{ "prim": "DUP", "annots": ["@alias_operation"] }, { "prim": "DUP" }, { "prim": "CAR", "annots": ["@address"] }, { "prim": "AMOUNT", "annots": ["@amount"] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@owned"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "NOT" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "CDR" }, { "prim": "CAR" }], [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@amount"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "COMPARE" }, { "prim": "LT" }, { "prim": "AND" }, { "prim": "IF", "args": [[[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "CDR" }, { "prim": "CAR" }], { "prim": "PUSH", "args": [{ "prim": "string" }, { "string": "bad amount, minimum = " }] }, { "prim": "PAIR" }, { "prim": "FAILWITH" }], [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "CAR" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@address"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "GET" }, { "prim": "IF_NONE", "args": [[{ "prim": "UNIT" }], [{ "prim": "DUP" }, [{ "prim": "CDR" }, { "prim": "CDR" }, { "prim": "CDR" }, { "prim": "CAR" }], { "prim": "IF_LEFT", "args": [[[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@address"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@sender"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "COMPARE" }, { "prim": "EQ" }, { "prim": "IF", "args": [[{ "prim": "UNIT" }], [{ "prim": "PUSH", "args": [{ "prim": "string" }, { "string": "can't change official alias" }] }, { "prim": "FAILWITH" }]] }, { "prim": "DIP", "args": [[{ "prim": "DROP" }]] }], [{ "prim": "IF_LEFT", "args": [[[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@owned"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "NOT" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@amount"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "CDR" }, { "prim": "CDR" }, { "prim": "CAR" }], { "prim": "COMPARE" }, { "prim": "GE" }, { "prim": "AND" }, [{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@contributor"] }]] }, { "prim": "SWAP" }], [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@sender"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "COMPARE" }, { "prim": "NEQ" }, { "prim": "AND" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@address"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@sender"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "COMPARE" }, { "prim": "NEQ" }, { "prim": "AND" }, { "prim": "IF", "args": [[[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }], [{ "prim": "CDR" }, { "prim": "CDR" }, { "prim": "CAR" }], { "prim": "PUSH", "args": [{ "prim": "string" }, { "string": "Bad amount, minimum = " }] }, { "prim": "PAIR" }, { "prim": "FAILWITH" }], [{ "prim": "UNIT" }]] }, { "prim": "DIP", "args": [[{ "prim": "DROP" }]] }], [{ "prim": "DROP" }, { "prim": "UNIT" }]] }]] }, { "prim": "DIP", "args": [[{ "prim": "DROP" }]] }]] }, { "prim": "DROP" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@storage"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "CDR" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "CAR" }, { "prim": "NOW" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@sender"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@address"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "COMPARE" }, { "prim": "EQ" }, { "prim": "IF", "args": [[{ "prim": "PUSH", "args": [{ "prim": "or", "args": [{ "prim": "unit" }, { "prim": "or", "args": [{ "prim": "address" }, { "prim": "unit" }] }] }, { "prim": "Left", "args": [{ "prim": "Unit" }] }] }], [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@owned"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "IF", "args": [[{ "prim": "PUSH", "args": [{ "prim": "or", "args": [{ "prim": "unit" }, { "prim": "or", "args": [{ "prim": "address" }, { "prim": "unit" }] }] }, { "prim": "Right", "args": [{ "prim": "Right", "args": [{ "prim": "Unit" }] }] }] }], [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@sender"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "LEFT", "args": [{ "prim": "unit" }] }, { "prim": "RIGHT", "args": [{ "prim": "unit" }] }]] }]] }, { "prim": "PAIR" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@amount"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "PAIR" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "CDR" }, { "prim": "CDR", "annots": ["@avatar"] }], { "prim": "PAIR" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "CDR" }, { "prim": "CAR", "annots": ["@alias"] }], { "prim": "PAIR", "annots": ["@alias"] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@address"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "DIP", "args": [[{ "prim": "SOME" }]] }, { "prim": "UPDATE" }, { "prim": "PAIR", "annots": ["@storage"] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@return"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "PAIR" }]] }, { "prim": "DIP", "args": [[{ "prim": "DROP" }, { "prim": "DROP" }, { "prim": "DROP" }]] }], [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@storage"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "NIL", "args": [{ "prim": "operation" }] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@lambda"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DROP" }]] }]] }]] }], [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "EXEC" }, { "prim": "CONS" }, { "prim": "PAIR" }]] }, { "prim": "DIP", "args": [[{ "prim": "DROP" }]] }], [{ "prim": "IF_LEFT", "args": [[[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@owned"] }]] }, { "prim": "SWAP" }], { "prim": "IF", "args": [[[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@storage"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "DUP" }, { "prim": "CAR" }, { "prim": "SWAP" }, { "prim": "CDR" }, { "prim": "CDR" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@min_fees"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "PAIR" }, { "prim": "SWAP" }, { "prim": "PAIR", "annots": ["@storage"] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@return"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "PAIR" }], [{ "prim": "PUSH", "args": [{ "prim": "string" }, { "string": "forbidden" }] }, { "prim": "FAILWITH" }]] }, { "prim": "DIP", "args": [[{ "prim": "DROP" }]] }], [{ "prim": "IF_LEFT", "args": [[{ "prim": "DUP" }, { "prim": "CAR", "annots": ["@address"] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@owned"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "IF", "args": [[[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@storage"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "DUP" }, { "prim": "CAR" }, { "prim": "SWAP" }, { "prim": "CDR" }, { "prim": "DUP" }, { "prim": "CAR" }, { "prim": "SWAP" }, { "prim": "CDR" }, { "prim": "CDR" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "CDR", "annots": ["@add"] }, { "prim": "IF", "args": [[[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "CDR" }, { "prim": "CDR" }, { "prim": "CAR" }], [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@address"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "DIP", "args": [[{ "prim": "PUSH", "args": [{ "prim": "bool" }, { "prim": "True" }] }]] }, { "prim": "UPDATE" }], [{ "prim": "PUSH", "args": [{ "prim": "nat" }, { "int": "1" }] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "CDR" }, { "prim": "CDR" }, { "prim": "CAR" }], { "prim": "SIZE" }, { "prim": "COMPARE" }, { "prim": "EQ" }, { "prim": "IF", "args": [[{ "prim": "PUSH", "args": [{ "prim": "string" }, { "string": "cannot remove last owner" }] }, { "prim": "FAILWITH" }], [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], [{ "prim": "CDR" }, { "prim": "CDR" }, { "prim": "CAR" }], [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@address"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "DIP", "args": [[{ "prim": "PUSH", "args": [{ "prim": "bool" }, { "prim": "False" }] }]] }, { "prim": "UPDATE" }]] }]] }, { "prim": "PAIR" }, { "prim": "SWAP" }, { "prim": "PAIR" }, { "prim": "SWAP" }, { "prim": "PAIR", "annots": ["@storage"] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "PAIR" }], [{ "prim": "PUSH", "args": [{ "prim": "string" }, { "string": "forbidden" }] }, { "prim": "FAILWITH" }]] }, { "prim": "DIP", "args": [[{ "prim": "DROP" }, { "prim": "DROP" }]] }], [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@owned"] }]] }, { "prim": "SWAP" }], { "prim": "IF", "args": [[[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@storage"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "DUP" }, { "prim": "CAR" }, { "prim": "SWAP" }, { "prim": "CDR" }, { "prim": "DUP" }, { "prim": "CAR" }, { "prim": "SWAP" }, { "prim": "CDR" }, { "prim": "DUP" }, { "prim": "CAR" }, { "prim": "SWAP" }, { "prim": "DROP" }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@lambda"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "SOME" }, { "prim": "SWAP" }, { "prim": "PAIR" }, { "prim": "SWAP" }, { "prim": "PAIR" }, { "prim": "SWAP" }, { "prim": "PAIR", "annots": ["@storage"] }, [{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DIP", "args": [[{ "prim": "DUP", "annots": ["@return"] }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }]] }, { "prim": "SWAP" }], { "prim": "PAIR" }], [{ "prim": "PUSH", "args": [{ "prim": "string" }, { "string": "forbidden" }] }, { "prim": "FAILWITH" }]] }, { "prim": "DIP", "args": [[{ "prim": "DROP" }]] }]] }]] }]] }, { "prim": "DIP", "args": [[{ "prim": "DROP" }, { "prim": "DROP" }, { "prim": "DROP" }, { "prim": "DROP" }, { "prim": "DROP" }]] }]] }], "storage": { "prim": "Pair", "args": [[], { "prim": "Pair", "args": [{ "int": "0" }, { "prim": "Pair", "args": [[{ "string": "tz1ZLyEcFfXyu75ASsjAt5hgaR4sU5PSbLkg" }], { "prim": "None" }] }] }] } }, "counter": "2" }
+
+export const storage = rpcContractResponse.script.code.find(
+    x => x.prim === 'storage'
+)!.args[0] as any;
+
+export const params = rpcContractResponse.script.code.find(
+    x => x.prim === 'parameter'
+)!.args[0];
