+++ packages/tezos-ts-michelson-encoder/src/tokens/comparable/key_hash.ts
@@ -1,0 +1,37 @@
+import { Token, TokenFactory, ComparableToken } from '../token';
+
+export class KeyHashToken extends Token implements ComparableToken {
+  static prim = 'key_hash';
+
+  constructor(
+    protected val: { prim: string; args: any[]; annots: any[] },
+    protected idx: number,
+    protected fac: TokenFactory
+  ) {
+    super(val, idx, fac);
+  }
+
+  public Execute(val: any): { [key: string]: any } {
+    return val.string;
+  }
+
+  public Encode(...args: any[]): any {
+    return { string: args[0] };
+  }
+
+  public ExtractSchema() {
+    return KeyHashToken.prim;
+  }
+
+  // tslint:disable-next-line: variable-name
+  public ToKey({ string }: any) {
+    return string;
+  }
+
+  public ToBigMapKey(val: string) {
+    return {
+      key: { string: val },
+      type: { prim: KeyHashToken.prim },
+    };
+  }
+}
