+++ packages/tezos-ts-michelson-encoder/src/tokens/lambda.ts
@@ -1,0 +1,32 @@
+import { Token, TokenFactory } from './token';
+
+export class LambdaToken extends Token {
+  static prim = 'lambda';
+
+  constructor(
+    protected val: { prim: string; args: any[]; annots: any[] },
+    protected idx: number,
+    protected fac: TokenFactory
+  ) {
+    super(val, idx, fac);
+  }
+
+  public Execute(val: any): { [key: string]: any } {
+    return val.string;
+  }
+
+  public Encode(...args: any[]): any {
+    return { string: args[0] };
+  }
+
+  public ExtractSchema() {
+    const leftToken = this.createToken(this.val.args[0], this.idx);
+    const rightToken = this.createToken(this.val.args[1], this.idx + 1);
+    return {
+      [LambdaToken.prim]: {
+        parameters: leftToken.ExtractSchema(),
+        returns: rightToken.ExtractSchema(),
+      },
+    };
+  }
+}
